+++ .storybook/main.js
@@ -19,6 +19,7 @@ module.exports = {
     "@storybook/addon-interactions",
     "@storybook/addon-a11y",
     "@storybook/addon-controls",
+    "storybook-dark-mode",
   ],
   framework: "@storybook/react",
   // we need to add aliases to webpack so it knows how to follow
@@ -28,9 +29,7 @@ module.exports = {
 
     return {
       ...baseConfig,
-      plugins: [
-        ...plugins,
-      ],
+      plugins: [...plugins],
       resolve: {
         ...resolve,
         alias: {