+++ .storybook/preview.js
@@ -1,7 +1,34 @@
 import "@nimbus-ds/styles/dist/index.css";
 
+import { themes } from "@storybook/theming";
+
+import { base, darkExtend } from "./theme";
+import ThemeProvider from "./ThemeProvider";
+import theme from "./theme";
+import "./style.css";
+
 export const parameters = {
+  darkMode: {
+    current: "light",
+    darkClass: "darkClass",
+    lightClass: "lightClass",
+    stylePreview: true,
+    light: {
+      ...themes.light,
+      ...base,
+    },
+    dark: {
+      ...darkExtend,
+      ...base,
+    },
+  },
+  backgrounds: {
+    disable: true,
+  },
   actions: { argTypesRegex: "^on[A-Z].*" },
+  docs: {
+    theme,
+  },
   controls: {
     matchers: {
       color: /(background|color)$/i,
@@ -9,3 +36,11 @@ export const parameters = {
     },
   },
 };
+
+export const decorators = [
+  (Story) => (
+    <ThemeProvider>
+      <Story />
+    </ThemeProvider>
+  ),
+];